Se quitan interfaces de los controladores rest, se quita el plugin de flyway y se pone como componente de boot, se quita la creacion manual del datasource y se deja la automatica
