+++ src/main/java/com/github/erodriguezg/springbootangular/config/WebConfig.java
@@ -4,20 +4,15 @@ import com.github.erodriguezg.javautils.CodecUtils;
 import com.github.erodriguezg.javautils.DateUtils;
 import com.github.erodriguezg.springbootangular.security.CorsFilter;
 import org.springframework.boot.autoconfigure.web.servlet.DispatcherServletAutoConfiguration;
-import org.springframework.boot.web.server.ErrorPage;
-import org.springframework.boot.web.server.ErrorPageRegistrar;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.context.request.RequestContextListener;
 import org.springframework.web.multipart.commons.CommonsMultipartResolver;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.DispatcherServlet;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.tuckey.web.filters.urlrewrite.UrlRewriteFilter;
 
 import javax.servlet.DispatcherType;
 import java.util.EnumSet;
@@ -29,22 +24,6 @@ import java.util.EnumSet;
 @EnableWebMvc
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
-
-    private static final String[] CLASSPATH_RESOURCE_LOCATIONS = {
-            "classpath:/META-INF/resources/", "classpath:/resources/",
-            "classpath:/static/", "classpath:/public/", "classpath:/frontend-dist/"};
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        if (!registry.hasMappingForPattern("/webjars/**")) {
-            registry.addResourceHandler("/webjars/**").addResourceLocations(
-                    "classpath:/META-INF/resources/webjars/");
-        }
-        if (!registry.hasMappingForPattern("/**")) {
-            registry.addResourceHandler("/**").addResourceLocations(
-                    CLASSPATH_RESOURCE_LOCATIONS);
-        }
-    }
 
     @Bean
     public CommonsMultipartResolver filterMultipartResolver() {
@@ -93,31 +72,6 @@ public class WebConfig implements WebMvcConfigurer {
         registration.setName("corsFilter");
         registration.setOrder(1);
         return registration;
-    }
-
-    /*
-    Filter URL REWRITE
-     */
-
-    @Bean
-    public FilterRegistrationBean<UrlRewriteFilter> urlRewriteFilter() {
-        FilterRegistrationBean<UrlRewriteFilter> registration = new FilterRegistrationBean<>();
-        registration.setFilter(new UrlRewriteFilter());
-        registration.addUrlPatterns("/*");
-        registration.setDispatcherTypes(EnumSet.of(DispatcherType.FORWARD, DispatcherType.REQUEST,
-                DispatcherType.INCLUDE, DispatcherType.ASYNC, DispatcherType.ERROR));
-        registration.setName("urlRewriteFilter");
-        registration.setOrder(1);
-        return registration;
-    }
-
-    @Bean
-    public ErrorPageRegistrar errorPageRegistrar() {
-        return container -> {
-            container.addErrorPages(new ErrorPage(HttpStatus.NOT_FOUND, "/404.html"));
-            container.addErrorPages(new ErrorPage(HttpStatus.UNAUTHORIZED, "/access.html"));
-            container.addErrorPages(new ErrorPage(HttpStatus.INTERNAL_SERVER_ERROR, "/error.html"));
-        };
     }
 
     /*
