--- src/main/java/com/github/erodriguezg/springbootangular/rest/ComunaRest.java
@@ -1,13 +1,0 @@
-package com.github.erodriguezg.springbootangular.rest;
-
-import com.github.erodriguezg.springbootangular.services.dto.ComunaDto;
-
-import java.util.List;
-
-public interface ComunaRest {
-
-    ComunaDto traerComunaPorIdComuna(Integer idComuna);
-
-    List<ComunaDto> traerComunasPorProvincia(int idProvincia);
-
-}
