--- src/main/java/com/github/erodriguezg/springbootangular/rest/HelloworldRest.java
@@ -1,9 +1,0 @@
-package com.github.erodriguezg.springbootangular.rest;
-
-public interface HelloworldRest {
-
-	String hello();
-
-	String helloProtected();
-
-}
