--- src/main/java/com/github/erodriguezg/springbootangular/rest/PerfilRest.java
@@ -1,11 +1,0 @@
-package com.github.erodriguezg.springbootangular.rest;
-
-import com.github.erodriguezg.springbootangular.services.dto.PerfilDto;
-
-import java.util.List;
-
-public interface PerfilRest {
-
-    List<PerfilDto> traerTodos();
-
-}
