--- src/main/java/com/github/erodriguezg/springbootangular/rest/RegionRest.java
@@ -1,11 +1,0 @@
-package com.github.erodriguezg.springbootangular.rest;
-
-import com.github.erodriguezg.springbootangular.services.dto.RegionDto;
-
-import java.util.List;
-
-public interface RegionRest {
-
-    List<RegionDto> traerTodas();
-
-}
