--- src/main/java/com/github/erodriguezg/springbootangular/rest/ShowcaseRest.java
@@ -1,13 +1,0 @@
-package com.github.erodriguezg.springbootangular.rest;
-
-public interface ShowcaseRest {
-
-    void lanzarErrorNegocio();
-
-    void lanzarErrorInterno();
-
-    void accionAdministrador();
-
-    void sleep();
-
-}
