+++ src/main/java/com/github/erodriguezg/springbootangular/rest/impl/RegionRest.java
@@ -1,8 +1,7 @@
 package com.github.erodriguezg.springbootangular.rest.impl;
 
-import com.github.erodriguezg.springbootangular.rest.PerfilRest;
-import com.github.erodriguezg.springbootangular.services.PerfilService;
-import com.github.erodriguezg.springbootangular.services.dto.PerfilDto;
+import com.github.erodriguezg.springbootangular.services.RegionService;
+import com.github.erodriguezg.springbootangular.services.dto.RegionDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,17 +11,16 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@RequestMapping("/perfiles")
-public class PerfilRestImpl implements PerfilRest {
+@RequestMapping("/regiones")
+public class RegionRest {
 
     @Autowired
-    private PerfilService perfilService;
+    private RegionService regionService;
 
-    @GetMapping("/todos")
+    @GetMapping("/todas")
     @PreAuthorize("isAuthenticated()")
-    @Override
-    public List<PerfilDto> traerTodos() {
-        return this.perfilService.traerTodos();
+    public List<RegionDto> traerTodas() {
+        return regionService.traerTodas();
     }
 
 }
