+++ src/main/java/com/github/erodriguezg/springbootangular/rest/impl/ShowcaseRest.java
@@ -1,6 +1,5 @@
 package com.github.erodriguezg.springbootangular.rest.impl;
 
-import com.github.erodriguezg.springbootangular.rest.ShowcaseRest;
 import com.github.erodriguezg.springbootangular.services.exceptions.LogicaNegocioException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,19 +10,17 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/showcase")
-public class ShowcaseRestImpl implements ShowcaseRest {
+public class ShowcaseRest {
 
-    private static final Logger log = LoggerFactory.getLogger(ShowcaseRestImpl.class);
+    private static final Logger log = LoggerFactory.getLogger(ShowcaseRest.class);
 
     @GetMapping("/error-negocio")
-    @Override
     public void lanzarErrorNegocio() {
         log.debug("--> lanzando error de negocio");
         throw new LogicaNegocioException("Hola, soy una excepción controlada en el flujo!");
     }
 
     @GetMapping("/error-interno")
-    @Override
     public void lanzarErrorInterno() {
         log.debug("--> lanzando error interno");
         throw new IllegalStateException("Hola, soy un bug!");
@@ -31,13 +28,11 @@ public class ShowcaseRestImpl implements ShowcaseRest {
 
     @GetMapping("/accion-admin")
     @PreAuthorize("hasAnyAuthority('Administrador')")
-    @Override
     public void accionAdministrador() {
         log.debug("--> si llegue aquí es porque tengo permiso de Administrador!");
     }
 
     @GetMapping("/sleep")
-    @Override
     public void sleep() {
         log.debug("-> comenzando (5 seg) zzZZzzzZzzzZZzzz");
         try {
