--- src/main/resources/application-QA.properties
@@ -1,13 +1,0 @@
-#BASE DATOS
-app.db.jdbcurl=jdbc:postgresql://localhost:5432/springboot-angular
-app.db.user=postgres
-app.db.password=postgres
-app.db.poolsize=10
-
-#JSON WEB TOKENS
-app.jwt.secret-phrase=frase-secreta-jwt-cambiame
-app.jwt.expiration-minutes=5
-
-#LOGGER
-logging.level.com.github.erodriguezg=DEBUG
-logging.level.org.hibernate.SQL=DEBUG
