+++ src/main/resources/db/migration/V2__CONTENIDO_TABLAS_TIPO.sql
