--- src/main/resources/public/404.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Spring Reactive Angular</title>
-</head>
-<body>
-
-<h1>Recurso no encontrado 404</h1>
-
-</body>
-</html>
