--- src/main/resources/public/inicio.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Spring Reactive Angular</title>
-</head>
-<body>
-
-<h1>Inicio</h1>
-
-</body>
-</html>
